Move useConditionalValueProp
