+++ ui/appui-react/src/appui-react/panel/DynamicPanel.tsx
@@ -10,10 +10,9 @@ import "./DynamicPanel.scss";
 import * as React from "react";
 import { Divider, IconButton } from "@itwin/itwinui-react";
 import { SvgCloseSmall } from "@itwin/itwinui-icons-react";
-import { useConditionalValue } from "../hooks/useConditionalValue.js";
-import type { ConditionalValue } from "../shared/ConditionalValue.js";
 import type { DynamicPanel, DynamicPanelSlice } from "./PanelsState.js";
 import { PanelSideContext } from "../layout/widget-panels/Panel.js";
+import { useConditionalValueProp } from "../shared/ConditionalValue.js";
 
 interface DynamicPanelProps {
   panel: DynamicPanel;
@@ -41,29 +40,6 @@ function DynamicPanelComponent(props: DynamicPanelProps) {
       <Divider className="uifw-panel-dynamicPanel_divider" />
       <div>{panel.content}</div>
     </div>
-  );
-}
-
-function isConditionalValue<T>(
-  value: T | ConditionalValue<T>
-): value is ConditionalValue<T> {
-  return (
-    typeof value === "object" &&
-    value !== null &&
-    "eventIds" in value &&
-    "getValue" in value
-  );
-}
-
-function useConditionalValueProp<T>(prop: T | ConditionalValue<T>): T {
-  return useConditionalValue(
-    () => {
-      if (isConditionalValue(prop)) {
-        return prop.getValue();
-      }
-      return prop;
-    },
-    isConditionalValue(prop) ? prop.eventIds : []
   );
 }
